2025/10/06
Added two version of latestGP.py
The first is same as before to test code one more time because
it sometimes depends on computer's performance when we run the code.
And the other one which is latestGP2.py is modified a little to
run code efficiently. Specifically I made this code with minimizing
operation 'np.mean' which usually takes a long time to run
just recycling already existing code.